Move tests for client.user_has_privilege() and associated tests into their own file
client.user_has_privilege() and client.user_can_view_unpublished_judgments()
are releated tests, so keep them together.
